Implemented click cli tool to make future updates easier
