+++ llm_dump/__init__.py
@@ -1,0 +1,5 @@
+"""
+LLM Dump Tool - A tool to dump various content sources into a single text file for LLM context
+"""
+
+__version__ = "0.1.0"
