+++ llm_dump/cli.py
@@ -1,0 +1,3 @@
+from llm_dump.group import cli
+from llm_dump.repo import git  # This registers the command
+from llm_dump.markdown import markdown  # This registers the command
