+++ llm_dump/group.py
@@ -1,0 +1,6 @@
+import click
+
+@click.group()
+def cli():
+    """Process various content sources for LLM context."""
+    pass
