+++ pyproject.toml
@@ -1,11 +1,12 @@
 [project]
 name = "llm-dump-tool"
-version = "0.1.0"
+version = "0.1.4"
 description = "A tool to dump various content sources into a single text file for LLM context"
 authors = [{ name = "Adrian Obleton", email = "obletonadrian@gmail.com" }]
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
+    "click>=8.1.7",
     "pathspec>=0.12.1",
     "pydantic>=2.10.1",
     "pytest>=8.3.4",
@@ -13,4 +14,4 @@ dependencies = [
 ]
 
 [project.scripts]
-llm-dump = "llm_dump:main"
+llm-dump = "llm_dump.cli:cli"
